Split diskcollectionpanel into a few files to make it more readable
