+++ src/ui/devices/disk/driveprops.ts
@@ -6,7 +6,6 @@ import { GoogleDrive } from "./googledrive"
 import { isHardDriveImage, RUN_MODE, MAX_DRIVES, replaceSuffix, FILE_SUFFIXES_DISK } from "../../../common/utility"
 // import { iconKey, iconData, iconName } from "../../img/iconfunctions"
 import { passSetDriveNewData, passSetDriveProps, passSetBinaryBlock, passPasteText, handleGetRunMode, passSetRunMode } from "../../main2worker"
-import { DISK_COLLECTION_ITEM_TYPE } from "../../diskdialog/diskcollectionpanel"
 import { showGlobalProgressModal } from "../../ui_utilities"
 import { internetArchiveUrlProtocol, getDiskImageUrlFromIdentifier } from "./internetarchive_utils"
 import { newReleases } from "./newreleases"
@@ -14,6 +13,7 @@ import { DiskBookmarks } from "./diskbookmarks"
 import { parseGameList } from "./totalreplayutilities"
 import { getHotReload, setHelpText } from "../../ui_settings"
 import { getDiskImageFromLocalStorage, setDiskImageToLocalStorage } from "../../localstorage"
+import { DISK_COLLECTION_ITEM_TYPE } from "../../diskdialog/diskpanel_utils"
 
 // Technically, all of these properties should be in the main2worker.ts file,
 // since they just maintain the state that needs to be passed to/from the
@@ -596,7 +596,7 @@ const checkForHelpFile = async (disk: string) => {
 }
 
 export const handleSetDiskFromFile = async (disk: string,
-  updateDisplay: UpdateDisplay, driveIndex: number = -1,
+  updateDisplay: UpdateDisplay | null, driveIndex: number = -1,
   callback?: (buffer: ArrayBuffer | null) => void) => {
   let data: ArrayBuffer
   try {
@@ -646,11 +646,11 @@ export const handleSetDiskFromFile = async (disk: string,
         if (helpFile.includes("Total%20Replay")) {
         helptext = parseGameList(helptext)
       }
-      updateDisplay(0, helptext)
+        updateDisplay?.(0, helptext)
       }      
     } catch {
       // If we don't have a help text file, just revert to the default text.
-      updateDisplay(0, "<Default>")
+      updateDisplay?.(0, "<Default>")
     }
   }
 }
