+++ src/ui/devices/disk/internetarchivedialog.tsx
@@ -6,7 +6,7 @@ import { faStar as faStarSolid } from "@fortawesome/free-solid-svg-icons"
 import { DiskBookmarks } from "./diskbookmarks"
 import { CLOUD_SYNC } from "../../../common/utility"
 import { svgInternetArchiveSoftware, svgInternetArchiveViews, svgInternetArchiveFavorites, svgInternetArchiveReviews, svgInternetArchiveLogo, svgInternetArchiveTitle } from "../../img/icon_internetarchive"
-import { DISK_COLLECTION_ITEM_TYPE } from "../../diskdialog/diskcollectionpanel"
+import { DISK_COLLECTION_ITEM_TYPE } from "../../diskdialog/diskpanel_utils"
 import { showGlobalProgressModal } from "../../ui_utilities"
 import { handleSetDiskFromURL } from "./driveprops"
 import { generateUrlFromInternetArchiveId } from "./internetarchive_utils"
