+++ src/ui/diskdialog/diskcollectionpanel.tsx
@@ -1,76 +1,28 @@
-import { ChangeEvent, useEffect, useLayoutEffect, useRef, useState } from "react"
+import { ChangeEvent, useEffect, useState } from "react"
 import "./diskcollectionpanel.css"
 import Flyout from "../flyout"
 import { faCommentDots, faCheckCircle, faClock, faCloud, faDownload, faFloppyDisk, faHardDrive, faStar } from "@fortawesome/free-solid-svg-icons"
-import { RUN_MODE } from "../../common/utility"
-import { buildProDosHdv, ImportedDiskFile, loadWozAndExtractProDosFiles, loadWozAndExtractDosImage, ensureDosVolumeHasHelloGreeting, PRODOS_FILE_TYPE_DOS_MASTER, PRODOS_FILE_TYPE_TEXT, MenuDiskEntry, classifyImageKind, determineVtocType, stripTwoImgHeader } from "../../common/prodos_hdv"
-import { loadAndConvertImageToHires } from "./screenshot_utils"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
 import { svgInternetArchiveLogo } from "../img/icon_internetarchive"
 import PopupMenu from "../controls/popupmenu"
 import { DISK_DRIVE_LABELS } from "../devices/disk/diskdrive"
-import { handleSetDiskFromCloudData, handleSetDiskFromFile, handleSetDiskFromURL } from "../devices/disk/driveprops"
-import { diskImages } from "../devices/disk/diskimages"
 import { newReleases } from "../devices/disk/newreleases"
 import { DiskBookmarks } from "../devices/disk/diskbookmarks"
 import {
-  addSessionVtocFailure,
   DiskCollectionSortMode,
   getPreferenceDiskCollectionSort,
-  getPreferenceNewReleasesChecked,
-  getPreferenceVtocType,
-  hasSessionVtocFailure,
   setPreferenceDiskCollectionSort,
   setPreferenceNewReleasesChecked,
-  setPreferenceVtocType
 } from "../localstorage"
 import { isMinimalTheme } from "../ui_settings"
-import { handleInputParams } from "../inputparams"
 import { faCircle } from "@fortawesome/free-regular-svg-icons"
 import { getDiskImageUrlFromIdentifier } from "../devices/disk/internetarchive_utils"
 import { showGlobalProgressModal } from "../ui_utilities"
-import { passSetRunMode } from "../main2worker"
-
-export enum DISK_COLLECTION_ITEM_TYPE {
-  A2TS_ARCHIVE,
-  INTERNET_ARCHIVE,
-  NEW_RELEASE,
-  CLOUD_DRIVE
-}
+import { sortDisks, DISK_COLLECTION_ITEM_TYPE, TAB_INDEX, getDiskCollection, getExportFilename, isDiskExportable, getExportBadgeInfo, loadDisk, createHdv } from "./diskpanel_utils"
+import { DiskItemTitle } from "./diskitemtitle"
+import { DiskPanelVtoc } from "./diskpanel_vtoc"
 
 const maxHdvBytes = 33554432
-
-// Recover the known byte size of a disk by matching its URL against the built-in
-// catalogs. Favorited disks are rebuilt from bookmarks that don't store a size,
-// so without this a favorited large disk (e.g. Total Replay) would report an
-// unknown size (-1) and slip past the export size limit.
-const getKnownFileSizeForUrl = (diskUrl?: string): number | undefined => {
-  if (!diskUrl) return undefined
-  const match = [...diskImages, ...newReleases].find(d => d.diskUrl?.toString() === diskUrl)
-  return match?.fileSize && match.fileSize > 0 ? match.fileSize : undefined
-}
-
-// Assigns a disk's VTOC type without downloading its bytes, when possible:
-//  - HDV images are hard-drive volumes that are always ProDOS, so their type is
-//    set directly by file extension (no download or decode needed).
-//  - Bookmarks persist their own VTOC type in their dbm- entry (the item is built
-//    with vtocType: diskBookmark.vtocType), so that stored value is authoritative
-//    and nothing further is restored here for them.
-//  - Other disks (built-in images, new releases) have no per-disk store, so a VTOC
-//    type determined in a previous session is restored from the permanent
-//    URL-keyed local-storage cache.
-// Anything still undefined is resolved later by the background effect, but only
-// if the disk is small enough to ever be exported.
-const restoreCachedVtocType = (item: DiskCollectionItem) => {
-  if (item.vtocType !== undefined) return
-  const url = item.diskUrl?.toString() || ""
-  if (url.toLowerCase().split(/[?#]/)[0].endsWith(".hdv")) {
-    item.vtocType = "prodos"
-    return
-  }
-  if (item.bookmarkId) return
-  item.vtocType = getPreferenceVtocType(url)
-}
 
 const minDate = new Date(0)
 const dateFormatter = new Intl.DateTimeFormat("en-US", {
@@ -78,48 +30,6 @@ const dateFormatter = new Intl.DateTimeFormat("en-US", {
   month: "numeric",
   day: "numeric"
 })
-
-type DownloadedExportDisk = {
-  item: DiskCollectionItem,
-  buffer: Uint8Array,
-  filename: string,
-}
-
-const getExportFilename = (diskCollectionItem: DiskCollectionItem, buffer: Uint8Array) => {
-  const fromCloud = diskCollectionItem.cloudData?.fileName?.trim()
-  if (fromCloud) {
-    return fromCloud
-  }
-
-  const rawName = diskCollectionItem.diskUrl.split("/").pop() || diskCollectionItem.title || "disk"
-  const decodedName = decodeURIComponent(rawName)
-  if (decodedName.includes(".")) {
-    return decodedName
-  }
-
-  const normalizedTitle = diskCollectionItem.title.replace(/[^A-Za-z0-9]+/g, "_").replace(/^_+|_+$/g, "") || "disk"
-  return `${normalizedTitle}.${buffer.byteLength <= 143360 ? "dsk" : "po"}`
-}
-
-const downloadExportHdv = (data: Uint8Array, filename: string) => {
-  const arrayBuffer = new ArrayBuffer(data.byteLength)
-  new Uint8Array(arrayBuffer).set(data)
-  const blob = new Blob([arrayBuffer], { type: "application/octet-stream" })
-  const url = URL.createObjectURL(blob)
-  const link = document.createElement("a")
-  link.href = url
-  link.download = filename
-  link.style.display = "none"
-  document.body.appendChild(link)
-  link.click()
-  document.body.removeChild(link)
-  URL.revokeObjectURL(url)
-}
-
-const TAB_INDEX_BUILT_IN = 0
-const TAB_INDEX_NEW_RELEASES = 1
-const TAB_INDEX_FAVORITES = 2
-const TAB_INDEX_SELECT = 3
 
 const diskCollectionSortOptions: Array<{ value: DiskCollectionSortMode; label: string }> = [
   { value: "name-asc", label: "A-Z" },
@@ -129,89 +39,18 @@ const diskCollectionSortOptions: Array<{ value: DiskCollectionSortMode; label: s
 ]
 
 const defaultSortModeByTab: Record<number, DiskCollectionSortMode> = {
-  [TAB_INDEX_BUILT_IN]: "name-asc",
-  [TAB_INDEX_NEW_RELEASES]: "date-newest",
-  [TAB_INDEX_FAVORITES]: "date-newest",
-  [TAB_INDEX_SELECT]: "name-asc",
+  [TAB_INDEX.BUILT_IN]: "name-asc",
+  [TAB_INDEX.NEW_RELEASES]: "date-newest",
+  [TAB_INDEX.FAVORITES]: "date-newest",
+  [TAB_INDEX.EXPORT]: "name-asc",
 }
 
-const getLastUpdatedMs = (item: DiskCollectionItem): number => {
-  if (!item.lastUpdated) return 0
-  const date = item.lastUpdated instanceof Date ? item.lastUpdated : new Date(item.lastUpdated)
-  const value = date.getTime()
-  return Number.isFinite(value) ? value : 0
-}
-
-const compareByNameAsc = (a: DiskCollectionItem, b: DiskCollectionItem): number => {
-  const byName = a.title.localeCompare(b.title)
-  if (byName !== 0) return byName
-  return getLastUpdatedMs(b) - getLastUpdatedMs(a)
-}
-
-const sortDisks = (disks: DiskCollectionItem[], mode: DiskCollectionSortMode): DiskCollectionItem[] => {
-  const sorted = [...disks]
-  sorted.sort((a, b) => {
-    switch (mode) {
-      case "name-desc": {
-        const byName = b.title.localeCompare(a.title)
-        if (byName !== 0) return byName
-        return getLastUpdatedMs(b) - getLastUpdatedMs(a)
-      }
-      case "date-newest": {
-        const byDate = getLastUpdatedMs(b) - getLastUpdatedMs(a)
-        if (byDate !== 0) return byDate
-        return compareByNameAsc(a, b)
-      }
-      case "date-oldest": {
-        const byDate = getLastUpdatedMs(a) - getLastUpdatedMs(b)
-        if (byDate !== 0) return byDate
-        return compareByNameAsc(a, b)
-      }
-      case "name-asc":
-      default:
-        return compareByNameAsc(a, b)
-    }
-  })
-  return sorted
+const getSortModeForTab = (tabIndex: number): DiskCollectionSortMode => {
+  return getPreferenceDiskCollectionSort(tabIndex) || defaultSortModeByTab[tabIndex] || "name-asc"
 }
 
 type DiskCollectionPanelProps = DisplayProps & {
   onDismissDialog?: () => void
-}
-
-// Renders a disk title that automatically shrinks its font size so the text
-// fits within the fixed-size title box instead of being clipped.
-const DiskItemTitle = (props: {
-  text: string,
-  className: string,
-  title: string,
-  onClick: (e: React.MouseEvent<HTMLDivElement>) => void,
-}) => {
-  const ref = useRef<HTMLDivElement>(null)
-
-  useLayoutEffect(() => {
-    const el = ref.current
-    if (!el) return
-    const fit = () => {
-      // Start from the CSS-defined size and only shrink if it overflows.
-      el.style.fontSize = ""
-      const maxSize = parseFloat(window.getComputedStyle(el).fontSize)
-      const minSize = 6
-      let size = maxSize
-      while (size > minSize && (el.scrollWidth > el.clientWidth || el.scrollHeight > el.clientHeight)) {
-        size -= 0.5
-        el.style.fontSize = `${size}px`
-      }
-    }
-    fit()
-    const observer = new ResizeObserver(fit)
-    observer.observe(el)
-    return () => observer.disconnect()
-  }, [props.text])
-
-  return (
-    <div ref={ref} className={props.className} title={props.title} onClick={props.onClick}>{props.text}</div>
-  )
 }
 
 const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
@@ -223,73 +62,20 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
   const [popupItem, setPopupItem] = useState<DiskCollectionItem>()
   const [activeTab, setActiveTab] = useState<number>(0)
   const [sortModeByTab, setSortModeByTab] = useState<Record<number, DiskCollectionSortMode>>(() => ({
-    [TAB_INDEX_BUILT_IN]: getPreferenceDiskCollectionSort(TAB_INDEX_BUILT_IN) || defaultSortModeByTab[TAB_INDEX_BUILT_IN],
-    [TAB_INDEX_NEW_RELEASES]: getPreferenceDiskCollectionSort(TAB_INDEX_NEW_RELEASES) || defaultSortModeByTab[TAB_INDEX_NEW_RELEASES],
-    [TAB_INDEX_FAVORITES]: getPreferenceDiskCollectionSort(TAB_INDEX_FAVORITES) || defaultSortModeByTab[TAB_INDEX_FAVORITES],
-    [TAB_INDEX_SELECT]: getPreferenceDiskCollectionSort(TAB_INDEX_SELECT) || defaultSortModeByTab[TAB_INDEX_SELECT],
+    [TAB_INDEX.BUILT_IN]: getSortModeForTab(TAB_INDEX.BUILT_IN),
+    [TAB_INDEX.NEW_RELEASES]: getSortModeForTab(TAB_INDEX.NEW_RELEASES),
+    [TAB_INDEX.FAVORITES]: getSortModeForTab(TAB_INDEX.FAVORITES),
+    [TAB_INDEX.EXPORT]: getSortModeForTab(TAB_INDEX.EXPORT),
   }))
   const [hasNewRelease, setHasNewRelease] = useState<boolean>(false)
   const [selectedDisks, setSelectedDisks] = useState<DiskCollectionItem[]>([])
   const [exportQueue, setExportQueue] = useState<DiskCollectionItem[]>([])
   const [downloadedDisks, setDownloadedDisks] = useState<DownloadedExportDisk[]>([])
 
-  // Tracks disk items whose VTOC download has been attempted during the current
-  // active-tab verification pass (success or failure), so we don't re-pick the
-  // same disk while iterating the queue. Cleared when the active tab changes.
-  const vtocResolveAttempted = useRef<Set<string>>(new Set())
-  // Bumped after a failed download so the verification effect advances to the
-  // next pending disk. CORS/network failures are retried when the user revisits
-  // the tab in a new browser session.
-  const [vtocCheckPass, setVtocCheckPass] = useState(0)
-  // Remembers which tab the current VTOC pass belongs to so a tab switch starts
-  // a fresh pass for the newly visible disks.
-  const vtocActiveTabRef = useRef<number | null>(null)
-  const vtocProgressVisibleRef = useRef(false)
 
-  // Stable identity for a disk across re-renders (bookmark id, cloud item id,
-  // disk URL, or finally the title).
-  const itemKey = (item: DiskCollectionItem): string =>
-    item.bookmarkId || item.cloudData?.itemId || item.diskUrl?.toString() || item.title
-
-  const getSortModeForTab = (tabIndex: number): DiskCollectionSortMode => {
-    return sortModeByTab[tabIndex] || defaultSortModeByTab[tabIndex] || "name-asc"
-  }
 
   if (isMinimalTheme()) {
     import("./diskcollectionpanel.minimal.css")
-  }
-
-  const isDiskExportable = (disk: DiskCollectionItem) => {
-    // A disk explicitly flagged exportDisabled is never exportable (e.g. whole-volume
-    // desktop OS boot disks that assume they own the boot volume root and can't run from
-    // an imported subdirectory). A disk whose VTOC has been determined to be neither DOS
-    // nor ProDOS ("other"), or a DOS 3.3 disk whose greeting installs a language-card DOS
-    // relocator ("dosup", which is incompatible with DOS.MASTER), cannot be exported. An
-    // undetermined (undefined) vtocType is treated as potentially exportable until
-    // background determination resolves it.
-    return !disk.exportDisabled && disk.fileSize < maxHdvBytes * 0.95 && disk.vtocType !== "other" && disk.vtocType !== "dosup"
-  }
-
-  // Determine the export badge state and tooltip for a disk. Exportable disks show
-  // a green badge; disks that cannot be exported show an orange badge explaining
-  // why; disks whose VTOC hasn't been determined yet show a yellow "pending" badge.
-  const getExportBadgeInfo = (disk: DiskCollectionItem): { state: "exportable" | "blocked" | "pending"; title: string } => {
-    if (disk.exportDisabled) {
-      return { state: "blocked", title: "Export is not currently supported for this disk" }
-    }
-    if (disk.fileSize >= maxHdvBytes * 0.95) {
-      return { state: "blocked", title: "Disk is too large to be exported" }
-    }
-    if (disk.vtocType === "other") {
-      return { state: "blocked", title: "Disk is not exportable due to copy protection" }
-    }
-    if (disk.vtocType === "dosup") {
-      return { state: "blocked", title: "Disk is not exportable due to DOS Master incompatibility" }
-    }
-    if (disk.vtocType === undefined) {
-      return { state: "pending", title: "Disk export status pending" }
-    }
-    return { state: "exportable", title: "Disk can be exported" }
   }
 
   const tabs = [
@@ -298,7 +84,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
       label: "Show Apple2TS collection",
       disks: sortDisks(
         diskCollection.filter(x => x.type == DISK_COLLECTION_ITEM_TYPE.A2TS_ARCHIVE),
-        getSortModeForTab(TAB_INDEX_BUILT_IN)
+        sortModeByTab[TAB_INDEX.BUILT_IN]
       ),
       isHighlighted: false
     },
@@ -307,7 +93,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
       label: "Show new releases",
       disks: sortDisks(
         diskCollection.filter(x => x.type == DISK_COLLECTION_ITEM_TYPE.NEW_RELEASE),
-        getSortModeForTab(TAB_INDEX_NEW_RELEASES)
+        sortModeByTab[TAB_INDEX.NEW_RELEASES]
       ),
       isHighlighted: hasNewRelease
     },
@@ -316,7 +102,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
       label: "Show favorites",
       disks: sortDisks(
         diskCollection.filter(x => x.type == DISK_COLLECTION_ITEM_TYPE.INTERNET_ARCHIVE || x.type == DISK_COLLECTION_ITEM_TYPE.CLOUD_DRIVE),
-        getSortModeForTab(TAB_INDEX_FAVORITES)
+        sortModeByTab[TAB_INDEX.FAVORITES]
       ),
       isHighlighted: false
     },
@@ -330,7 +116,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
       // too large are excluded.
       disks: sortDisks(
         diskCollection.filter(x => x.vtocType !== undefined && isDiskExportable(x)),
-        getSortModeForTab(TAB_INDEX_SELECT)
+        sortModeByTab[TAB_INDEX.EXPORT]
       ),
       isHighlighted: false
     }
@@ -351,76 +137,8 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
     return false
   }
 
-  const loadDisk = (
-    driveIndex: number,
-    diskCollectionItem: DiskCollectionItem | undefined = popupItem,
-    callback?: (buffer: ArrayBuffer | null) => void) => {
-    // We want to restart the emulator when loading a disk from our Disk Collection
-    passSetRunMode(RUN_MODE.IDLE)
-
-    if (diskCollectionItem) {
-      if (diskCollectionItem.type == DISK_COLLECTION_ITEM_TYPE.CLOUD_DRIVE && diskCollectionItem.cloudData) {
-        handleSetDiskFromCloudData(diskCollectionItem.cloudData, driveIndex, callback)
-      } else if (diskCollectionItem.diskUrl && !diskCollectionItem.diskUrl.includes("://")) {
-        handleSetDiskFromFile(diskCollectionItem.diskUrl, props.updateDisplay, driveIndex, callback)
-      } else {
-        handleSetDiskFromURL(diskCollectionItem.diskUrl || "", undefined, driveIndex, diskCollectionItem.cloudData, callback)
-      }
-      if (diskCollectionItem.params && !callback) {
-        handleInputParams(diskCollectionItem.params)
-      }
-    }
-
-    if (!callback) {
-      setIsFlyoutOpen(false)
-    }
-  }
-
-  // Downloads a disk's bytes without disturbing the running emulator. Unlike
-  // loadDisk(), this never changes the run mode or loads the disk into a drive;
-  // it simply resolves with the raw buffer (or null on failure) via callback.
-  // This runs as silent background verification (to populate export badges), so
-  // it deliberately does NOT show a blocking progress modal.
-  const fetchDiskBufferForItem = (diskCollectionItem: DiskCollectionItem): Promise<Uint8Array | null> => {
-    return new Promise((resolve) => {
-      const cb = (buffer: ArrayBuffer | null) => resolve(buffer ? new Uint8Array(buffer) : null)
-      if (diskCollectionItem.type == DISK_COLLECTION_ITEM_TYPE.CLOUD_DRIVE && diskCollectionItem.cloudData) {
-        handleSetDiskFromCloudData(diskCollectionItem.cloudData, -1, cb)
-      } else if (diskCollectionItem.diskUrl && !diskCollectionItem.diskUrl.includes("://")) {
-        handleSetDiskFromFile(diskCollectionItem.diskUrl, props.updateDisplay, -1, cb)
-      } else {
-        handleSetDiskFromURL(diskCollectionItem.diskUrl || "", undefined, -1, diskCollectionItem.cloudData, cb)
-      }
-    })
-  }
-
-  // Stores a determined VTOC type onto the in-memory collection item and persists
-  // it so it only needs to be determined once. Bookmarks keep their type in their
-  // own dbm- entry; other disks (built-in images, new releases) use the shared
-  // URL-keyed cache. A disk is never written to both, so a bookmark can't pick up
-  // a stale URL-cache value (and vice versa).
-  const persistVtocType = (diskCollectionItem: DiskCollectionItem, vtocType: VtocType) => {
-    diskCollectionItem.vtocType = vtocType
-
-    if (diskCollectionItem.bookmarkId) {
-      const bookmark = diskBookmarks.get(diskCollectionItem.bookmarkId)
-      if (bookmark) {
-        bookmark.vtocType = vtocType
-        diskBookmarks.set(bookmark)
-      }
-    } else {
-      // A disk's VTOC type never changes, so cache it by URL. This lets
-      // non-bookmarked disks (e.g. new releases) avoid re-downloading their bytes
-      // to redetermine the VTOC on every visit.
-      setPreferenceVtocType(diskCollectionItem.diskUrl.toString(), vtocType)
-    }
-
-    // Trigger a re-render so the export filter reflects the new VTOC type.
-    setDiskCollection((prev) => [...prev])
-  }
-
   const handleItemRightClick = (diskCollectionItem: DiskCollectionItem) => (event: React.MouseEvent<HTMLElement>) => {
-    if (activeTab == TAB_INDEX_SELECT) {
+    if (activeTab == TAB_INDEX.EXPORT) {
       setSelectPopupLocation([event.clientX, event.clientY])
     } else {
       setPopupItem(diskCollectionItem)
@@ -504,7 +222,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
   }
 
   const handleExportClick = () => {
-    const newExportQueue = tabs[TAB_INDEX_SELECT].disks.filter((diskCollectionItem) => {
+    const newExportQueue = tabs[TAB_INDEX.EXPORT].disks.filter((diskCollectionItem) => {
       return selectedDisks.includes(diskCollectionItem) && isDiskExportable(diskCollectionItem)
     })
 
@@ -541,7 +259,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
       setExportQueue((prevExportQueue) => prevExportQueue.slice(1))
     } else if (exportQueue.length > 0) {
       showGlobalProgressModal(true, `Fetching disk ${selectedDisks.length - exportQueue.length + 1}/${selectedDisks.length}`)
-      loadDisk(-1, exportQueue[0], processExportQueue)
+      loadDisk(-1, exportQueue[0], props.updateDisplay, processExportQueue)
     }
   }
 
@@ -569,166 +287,13 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
     return hdvSize <= 0 || hdvSize > maxHdvBytes
   }
 
-  const createHdv = async () => {
-    if (downloadedDisks.length === 0) {
-      return
-    }
-
-    const orderedDownloadedDisks = tabs[TAB_INDEX_SELECT].disks
-      .filter((diskCollectionItem) => selectedDisks.includes(diskCollectionItem))
-      .map((diskCollectionItem) => downloadedDisks.find((downloadedDisk) => downloadedDisk.item === diskCollectionItem))
-      .filter((downloadedDisk): downloadedDisk is DownloadedExportDisk => downloadedDisk !== undefined)
-
-    showGlobalProgressModal(true, "Creating HDV image")
-    try {
-      const wozExtractedByIndex = new Map<number, ImportedDiskFile[]>()
-      // WOZ DOS disks decoded to a flat DOS 3.3 (.dsk) sector image, keyed by disk index.
-      // DOS.MASTER runtime volumes must be plain DOS-order sector images, but classifyImageKind
-      // (unlike determineVtocType used for export eligibility) does not decode the WOZ
-      // bitstream, so decode it here to keep detection and writing consistent.
-      const wozDosImageByIndex = new Map<number, Uint8Array>()
-      for (let i = 0; i < orderedDownloadedDisks.length; i++) {
-        const disk = orderedDownloadedDisks[i]
-        if (!disk.filename.toLowerCase().endsWith(".woz")) continue
-        const extracted = loadWozAndExtractProDosFiles(disk.buffer)
-        if (extracted.length > 0) {
-          wozExtractedByIndex.set(i, extracted)
-          continue
-        }
-        const dosImage = loadWozAndExtractDosImage(disk.buffer)
-        if (dosImage) {
-          wozDosImageByIndex.set(i, dosImage)
-        }
-      }
-
-      const fileKinds = orderedDownloadedDisks.map((downloadedDisk) =>
-        classifyImageKind(downloadedDisk.filename, downloadedDisk.buffer)
-      )
-
-      for (const [index] of wozExtractedByIndex) {
-        fileKinds[index] = "prodos"
-      }
-
-      for (const [index] of wozDosImageByIndex) {
-        fileKinds[index] = "dos"
-      }
-
-      const fileEntries = orderedDownloadedDisks.map((downloadedDisk, index) => {
-        const fileKind = fileKinds[index]
-        const isText = downloadedDisk.filename.toUpperCase().endsWith(".TXT")
-        // Prefer the decoded DOS-order sector image for WOZ DOS disks so the runtime
-        // volume is a plain .dsk that DOS.MASTER can read; otherwise use the raw buffer.
-        // Strip any 2IMG (.2mg) 64-byte header so the ProDOS/DOS block image parses from
-        // block 0 (otherwise extraction finds no files and the disk only shows
-        // "PRODOS FILES IMPORTED"/CATALOG instead of launching).
-        const wozDosImage = wozDosImageByIndex.get(index)
-        let data = wozDosImage ?? stripTwoImgHeader(downloadedDisk.buffer)
-        if (fileKind === "dos" && !isText) {
-          // DOS.MASTER always runs a greeting named HELLO on the selected volume; ensure
-          // one exists (chaining to the source disk's real greeting) so booting a volume
-          // whose greeting was not named HELLO does not fail with FILE NOT FOUND.
-          const greeting = ensureDosVolumeHasHelloGreeting(data)
-          data = greeting.image
-          if (greeting.action === "injected") {
-            console.log(`[HDV Export] Injected HELLO greeting into ${downloadedDisk.filename}: ${greeting.command}`)
-          }
-        }
-        return {
-          name: downloadedDisk.filename.split(".")[0].slice(0, 15),
-          type: isText
-            ? PRODOS_FILE_TYPE_TEXT
-            : (fileKind === "dos" ? PRODOS_FILE_TYPE_DOS_MASTER : 0xE0),
-          data,
-          auxType: 0x0000,
-        }
-      })
-
-      // Convert each disk's preview screenshot to Apple II hi-res, stamping the
-      // Apple2TS logo into the corner. Conversion is deterministic; the source image is
-      // fetched through a localStorage cache (see resolveScreenshotUrlWithCache) so each
-      // screenshot only hits the network/app assets once across exports.
-      const screenshots: Array<{ name: string; data: Uint8Array | null }> = []
-      for (let index = 0; index < orderedDownloadedDisks.length; index++) {
-        const disk = orderedDownloadedDisks[index]
-        const screenshotData = await loadAndConvertImageToHires(disk.item.imageUrl, true, index + 1)
-        screenshots.push({
-          name: disk.filename.split(".")[0].slice(0, 15),
-          data: screenshotData,
-        })
-      }
-
-      // Create menu metadata with screenshot references
-      const menuEntries: MenuDiskEntry[] = orderedDownloadedDisks.map((disk, index) => ({
-        filename: disk.filename.split(".")[0].slice(0, 15),
-        sourceFilename: disk.filename,
-        displayName: disk.item.title,
-        screenshotData: screenshots[index].data || undefined,
-        imageKind: fileKinds[index],
-        wozExtractedProDosFiles: wozExtractedByIndex.get(index),
-      }))
-
-      const hdvData = await buildProDosHdv(fileEntries, "APPLE2TS", undefined, menuEntries)
-      downloadExportHdv(hdvData, "APPLE2TS.HDV")
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error)
-      alert(`Failed to build ProDOS HDV: ${message}`)
-    } finally {
-      showGlobalProgressModal(false)
-      setDownloadedDisks([])
-      setSelectedDisks([])
-      setActiveTab(0)
-      setIsFlyoutOpen(false)
-      props.onDismissDialog?.()
-    }
-  }
-
   useEffect(() => {
     setDiskBookmarks(new DiskBookmarks())
   }, [isFlyoutOpen])
 
   useEffect(() => {
-    const newDiskCollection: DiskCollectionItem[] = []
-    const newReleasesChecked = new Date(getPreferenceNewReleasesChecked())
-
-    // Load built-in disk images
-    diskImages.forEach((diskImage) => {
-      diskImage.type = DISK_COLLECTION_ITEM_TYPE.A2TS_ARCHIVE
-      restoreCachedVtocType(diskImage)
-      newDiskCollection.push(diskImage)
-    })
-
-    // Load favorites
-    for (const diskBookmark of diskBookmarks) {
-      const item: DiskCollectionItem = {
-        type: diskBookmark.type,
-        title: diskBookmark.title,
-        lastUpdated: new Date(diskBookmark.lastUpdated),
-        diskUrl: diskBookmark.diskUrl ? diskBookmark.diskUrl : "",
-        imageUrl: diskBookmark.screenshotUrl?.toString(),
-        detailsUrl: diskBookmark.cloudData?.detailsUrl ? diskBookmark.cloudData.detailsUrl : diskBookmark.detailsUrl?.toString(),
-        bookmarkId: diskBookmark.id,
-        cloudData: diskBookmark.cloudData,
-        fileSize: diskBookmark.cloudData?.fileSize || getKnownFileSizeForUrl(diskBookmark.diskUrl?.toString()) || -1,
-        vtocType: diskBookmark.vtocType
-      }
-      restoreCachedVtocType(item)
-      newDiskCollection.push(item)
-    }
-
-    // Load new releases
-    newReleases.forEach((newRelease) => {
-      newRelease.type = DISK_COLLECTION_ITEM_TYPE.NEW_RELEASE
-      // Reuse the previously determined VTOC type (or HDV assumption) so we don't
-      // re-download the disk's bytes just to redetermine an unchanging value.
-      restoreCachedVtocType(newRelease)
-      newDiskCollection.push(newRelease)
-
-      if (newRelease.lastUpdated >= newReleasesChecked) {
-        setHasNewRelease(true)
-      }
-    })
-
-    setDiskCollection(newDiskCollection)
+    const newDiskCollectionFinal = getDiskCollection(diskBookmarks, newReleases)
+    setDiskCollection(newDiskCollectionFinal)
   }, [diskBookmarks])
 
   useEffect(() => {
@@ -736,139 +301,19 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
     if (exportQueue.length > 0) {
       processExportQueue()
     } else if (downloadedDisks.length > 0) {
-      createHdv()
+      const orderedDownloadedDisks = tabs[TAB_INDEX.EXPORT].disks
+        .filter((diskCollectionItem) => selectedDisks.includes(diskCollectionItem))
+        .map((diskCollectionItem) => downloadedDisks.find((downloadedDisk) => downloadedDisk.item === diskCollectionItem))
+        .filter((downloadedDisk): downloadedDisk is DownloadedExportDisk => downloadedDisk !== undefined)
+      createHdv(orderedDownloadedDisks)
+      setDownloadedDisks([])
+      setSelectedDisks([])
+      setActiveTab(0)
+      setIsFlyoutOpen(false)
+      props.onDismissDialog?.()
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [exportQueue])
-
-  // While the Export tab is visible, fill in (and cache) the VTOC type of any
-  // disk that doesn't already have one by downloading its bytes. This keeps VTOC
-  // probing and its progress modal scoped to export workflows only. Built-in
-  // disks, new releases, and bookmarks are treated identically: an HDV or
-  // previously-cached type is filled in without a download (see
-  // restoreCachedVtocType), and only disks small enough to ever be exported are
-  // resolved here -- so large disks and un-downloadable (e.g. CORS-blocked)
-  // disks never trigger a download and never show as exportable. Disks are
-  // resolved one at a time to avoid a download stampede, and each result is
-  // cached in local storage so a given disk is only ever downloaded once.
-  // Download failures are remembered for the browser session (sessionStorage) so
-  // they aren't re-attempted on reload; they are retried in a new browser
-  // session.
-  useEffect(() => {
-    const shouldShowVtocProgress = activeTab === TAB_INDEX_SELECT
-    // The panel content is shown when the flyout is open (minimal theme) or
-    // always (classic theme renders it inside a dialog), matching Flyout's own
-    // render condition. Gate verification on actual visibility so it runs in
-    // both themes, not just when isFlyoutOpen is toggled.
-    const panelVisible = isFlyoutOpen || !isMinimalTheme()
-    if (!panelVisible) {
-      if (vtocProgressVisibleRef.current) {
-        showGlobalProgressModal(false)
-        vtocProgressVisibleRef.current = false
-      }
-      vtocActiveTabRef.current = null
-      return
-    }
-
-    // While exporting or building, progress ownership belongs to the export
-    // pipeline; avoid overlapping VTOC progress updates.
-    if (exportQueue.length > 0 || downloadedDisks.length > 0) {
-      if (vtocProgressVisibleRef.current) {
-        showGlobalProgressModal(false)
-        vtocProgressVisibleRef.current = false
-      }
-      return
-    }
-
-    // On a tab change, start a fresh pass for the newly visible disks. Disks
-    // that already resolved keep their cached vtocType and are not re-picked.
-    if (vtocActiveTabRef.current !== activeTab) {
-      vtocActiveTabRef.current = activeTab
-      vtocResolveAttempted.current.clear()
-    }
-
-    // Only resolve disks that are visible in the current tab, so we don't
-    // background-download disks the user can't see. The Export tab is the
-    // superset of all exportable disks, but its rendered list is pre-filtered to
-    // already-determined disks; use the full exportable candidate set for it so
-    // its not-yet-determined disks still get resolved when it's navigated to.
-    const visibleCandidates = activeTab === TAB_INDEX_SELECT
-      ? diskCollection.filter(isDiskExportable)
-      : tabs[activeTab].disks.filter(isDiskExportable)
-    const pendingCandidates = visibleCandidates.filter((item) =>
-      item.vtocType === undefined &&
-      isDiskExportable(item) &&
-      !hasSessionVtocFailure(item.diskUrl.toString())
-    )
-    const pending = pendingCandidates.find((item) =>
-      !vtocResolveAttempted.current.has(itemKey(item))
-    )
-    if (!pending) {
-      if (vtocProgressVisibleRef.current) {
-        showGlobalProgressModal(false)
-        vtocProgressVisibleRef.current = false
-      }
-      return
-    }
-
-    const currentDisk = vtocResolveAttempted.current.size + 1
-    const totalDisks = vtocResolveAttempted.current.size + pendingCandidates.length
-    if (shouldShowVtocProgress) {
-      showGlobalProgressModal(true, `Fetching disk metadata ${currentDisk}/${totalDisks}`)
-      vtocProgressVisibleRef.current = true
-    } else if (vtocProgressVisibleRef.current) {
-      showGlobalProgressModal(false)
-      vtocProgressVisibleRef.current = false
-    }
-
-    vtocResolveAttempted.current.add(itemKey(pending))
-    // Capture the ref's Set (stable across renders) and the item key so the
-    // cleanup below doesn't read a ref during teardown (which lint flags).
-    const attempted = vtocResolveAttempted.current
-    const pendingKey = itemKey(pending)
-    let cancelled = false
-    let settled = false
-
-    fetchDiskBufferForItem(pending)
-      .then((data) => {
-        settled = true
-        if (cancelled) {
-          return
-        }
-        if (!data) {
-          // Download failed (CORS/network). Remember the failure for this browser
-          // session so we don't re-attempt the download on every reload, then
-          // advance to the next disk. It will be retried in a new session.
-          addSessionVtocFailure(pending.diskUrl.toString())
-          setVtocCheckPass((pass) => pass + 1)
-          return
-        }
-        const filename = getExportFilename(pending, data)
-        // Cache the determined VTOC (and persist it for bookmarks).
-        persistVtocType(pending, determineVtocType(filename, data))
-      })
-      .catch(() => {
-        settled = true
-        if (!cancelled) {
-          addSessionVtocFailure(pending.diskUrl.toString())
-          setVtocCheckPass((pass) => pass + 1)
-        }
-      })
-
-    return () => {
-      cancelled = true
-      // If this effect re-ran (e.g. diskCollection changed during the panel-open
-      // re-render storm) before the fetch settled, un-mark the disk so it is
-      // retried on the next pass. Otherwise it would be stranded with an
-      // undefined VTOC -- showing a non-exportable badge until the panel is
-      // reopened. This most often struck the first disk, whose fetch is in flight
-      // exactly while the collection is still settling.
-      if (!settled) {
-        attempted.delete(pendingKey)
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [activeTab, isFlyoutOpen, diskCollection, vtocCheckPass, exportQueue.length, downloadedDisks.length])
 
   return (
     <Flyout
@@ -895,7 +340,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
       <div className="disk-collection-panel"
         onClick={(e) => { if (e.target === e.currentTarget) e.stopPropagation() }}>
         {tabs[activeTab].disks.map((diskCollectionItem, index) => {
-          const isExportTab = activeTab == TAB_INDEX_SELECT
+          const isExportTab = activeTab == TAB_INDEX.EXPORT
           const isDisabledForExport = isExportTab && !isDiskExportable(diskCollectionItem)
 
           return (
@@ -903,7 +348,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
               key={`dcp-${tabs[activeTab].label}-${index}`}
               className={`dcp-item${isDisabledForExport ? " dcp-item-disabled" : ""}`}
               onClick={(e) => {
-                if (activeTab == TAB_INDEX_SELECT && !isDisabledForExport) {
+                if (activeTab == TAB_INDEX.EXPORT && !isDisabledForExport) {
                   toggleSelectedItem(diskCollectionItem)
                   e.stopPropagation()
                 }
@@ -914,7 +359,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
                   className={`dcp-item-title ${diskCollectionItem.detailsUrl ? "dcp-item-title-link" : ""}`}
                   title={diskCollectionItem.detailsUrl ? `Click to show details for "${diskCollectionItem.title}"` : diskCollectionItem.title}
                   onClick={(e) => {
-                    if (activeTab != TAB_INDEX_SELECT && diskCollectionItem.detailsUrl) {
+                    if (activeTab != TAB_INDEX.EXPORT && diskCollectionItem.detailsUrl) {
                       handleHelpClick(diskCollectionItem)(e as React.MouseEvent<HTMLElement>)
                     }
                   }} />
@@ -924,8 +369,9 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
                 className="dcp-item-image-box"
                 title={`Click to load disk "${diskCollectionItem.title}"`}
                 onClick={() => {
-                  if (activeTab != TAB_INDEX_SELECT) {
-                    loadDisk(-1, diskCollectionItem)
+                  if (activeTab != TAB_INDEX.EXPORT) {
+                    loadDisk(-1, diskCollectionItem, props.updateDisplay)
+                    setIsFlyoutOpen(false)
                   }
                 }}
                 onContextMenu={handleItemRightClick(diskCollectionItem)}
@@ -939,7 +385,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
                       onClick={handleBookmarkClick(diskCollectionItem)}>
                       <FontAwesomeIcon icon={faStar} size="lg" className="dcp-item-bookmark-icon" />
                     </div>}
-                  {activeTab == TAB_INDEX_SELECT && isDiskExportable(diskCollectionItem) &&
+                  {activeTab == TAB_INDEX.EXPORT && isDiskExportable(diskCollectionItem) &&
                     <div
                       className="dcp-item-select"
                       title={`Click to ${selectedDisks.includes(diskCollectionItem) ? "remove to" : "add from"} selected disks`}
@@ -954,7 +400,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
                     {diskCollectionItem.type == DISK_COLLECTION_ITEM_TYPE.NEW_RELEASE &&
                       <div className="dcp-item-new" title="Disk is a new release">
                         <FontAwesomeIcon icon={faClock} size="lg" className="dcp-item-new-icon" onClick={(event) => {
-                          if (activeTab != TAB_INDEX_SELECT) {
+                          if (activeTab != TAB_INDEX.EXPORT) {
                             event.stopPropagation()
                           }
                         }} />
@@ -963,7 +409,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
                     {diskCollectionItem.type == DISK_COLLECTION_ITEM_TYPE.A2TS_ARCHIVE &&
                       <div className="dcp-item-a2ts" title="Disk is part of the Apple2TS collection">
                         <FontAwesomeIcon icon={faFloppyDisk} size="lg" className="dcp-item-a2ts-icon" onClick={(event) => {
-                          if (activeTab != TAB_INDEX_SELECT) {
+                          if (activeTab != TAB_INDEX.EXPORT) {
                             event.stopPropagation()
                           }
                         }} />
@@ -974,7 +420,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
                         <svg
                           className="dcp-item-ia-icon"
                           onClick={(event) => {
-                            if (activeTab != TAB_INDEX_SELECT) {
+                            if (activeTab != TAB_INDEX.EXPORT) {
                               event.stopPropagation()
                             }
                           }}
@@ -985,7 +431,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
                     {diskCollectionItem.type == DISK_COLLECTION_ITEM_TYPE.CLOUD_DRIVE &&
                       <div className="dcp-item-cloud" title={`Disk is synced via ${diskCollectionItem.cloudData?.providerName}`}>
                         <FontAwesomeIcon icon={faCloud} size="lg" className="dcp-item-cloud-icon" onClick={(event) => {
-                          if (activeTab != TAB_INDEX_SELECT) {
+                          if (activeTab != TAB_INDEX.EXPORT) {
                             event.stopPropagation()
                           }
                         }} />
@@ -1002,7 +448,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
                           className={`dcp-item-export-badge${badgeStateClass}`}
                           title={exportBadge.title}>
                           <FontAwesomeIcon icon={faDownload} size="lg" className="dcp-item-export-badge-icon" onClick={(event) => {
-                            if (activeTab != TAB_INDEX_SELECT) {
+                            if (activeTab != TAB_INDEX.EXPORT) {
                               event.stopPropagation()
                             }
                           }} />
@@ -1012,7 +458,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
                     })()}
                   </div>
                   <div className="dcp-item-icon-right-group">
-                    {activeTab == TAB_INDEX_SELECT &&
+                    {activeTab == TAB_INDEX.EXPORT &&
                       <div
                         className="dcp-item-report"
                         title="Report an export issue"
@@ -1039,7 +485,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
           <select
             id="dcp-sort-select"
             className="dcp-sort-select"
-            value={getSortModeForTab(activeTab)}
+            value={sortModeByTab[activeTab]}
             onChange={handleSortModeChange}
             onClick={(e) => e.stopPropagation()}>
             {diskCollectionSortOptions.map((option) => (
@@ -1048,17 +494,17 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
           </select>
         </div>
         <div className="dcp-export-size-cell">
-          {activeTab == TAB_INDEX_SELECT &&
+          {activeTab == TAB_INDEX.EXPORT &&
             <div className="dcp-export-size"><span className={`${estimateHdvSize() > maxHdvBytes ? "dcp-export-size-exceeded" : ""}`}>{formatBytes(estimateHdvSize())}</span> / {formatBytes(maxHdvBytes)}</div>
           }
         </div>
-        <div className={`dcp-export-controls${activeTab == TAB_INDEX_SELECT ? "" : " dcp-export-row-inactive"}`}>
+        <div className={`dcp-export-controls${activeTab == TAB_INDEX.EXPORT ? "" : " dcp-export-row-inactive"}`}>
           <button
             className="dcp-export-button"
-            disabled={activeTab != TAB_INDEX_SELECT || isExportButtonDisabled()}
+            disabled={activeTab != TAB_INDEX.EXPORT || isExportButtonDisabled()}
             onClick={(e) => {
               e.stopPropagation()
-              if (activeTab == TAB_INDEX_SELECT) {
+              if (activeTab == TAB_INDEX.EXPORT) {
                 handleExportClick()
               }
             }}>Export</button>
@@ -1081,7 +527,8 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
               isSelected: () => { return false },
               onClick: () => {
                 setDrivePopupLocation(undefined)
-                loadDisk(i, popupItem)
+                loadDisk(i, popupItem, props.updateDisplay)
+                setIsFlyoutOpen(false)
               }
             }
           )),
@@ -1093,7 +540,8 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
               isSelected: () => { return false },
               onClick: () => {
                 setDrivePopupLocation(undefined)
-                loadDisk(i, popupItem)
+                loadDisk(i, popupItem, props.updateDisplay)
+                setIsFlyoutOpen(false)
               }
             }
           ))
@@ -1114,7 +562,7 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
             icon: faCheckCircle,
             onClick: async () => {
               const newSelectedDisks = [...selectedDisks]
-              for (const diskCollectionItem of tabs[TAB_INDEX_SELECT].disks) {
+              for (const diskCollectionItem of tabs[TAB_INDEX.EXPORT].disks) {
                 if (isDiskExportable(diskCollectionItem) && !newSelectedDisks.includes(diskCollectionItem)) {
                   await prepareSelectedItem(diskCollectionItem)
                   newSelectedDisks.push(diskCollectionItem)
@@ -1131,6 +579,17 @@ const DiskCollectionPanel = (props: DiskCollectionPanelProps) => {
             }
           }
         ]]}
+      />
+      <DiskPanelVtoc
+        activeTab={activeTab}
+        isFlyoutOpen={isFlyoutOpen}
+        diskBookmarks={diskBookmarks}
+        setDiskCollection={setDiskCollection}
+        exportQueue={exportQueue}
+        downloadedDisks={downloadedDisks}
+        visibleCandidates={activeTab === TAB_INDEX.EXPORT
+          ? diskCollection.filter(isDiskExportable)
+          : tabs[activeTab].disks.filter(isDiskExportable)}
       />
     </Flyout >
   )
